+++ elfi/methods/bo/gpy_regression.py
